+++ lib/ImageType.php
@@ -1,0 +1,13 @@
+<?php
+
+class ImageType
+{
+    const SELFIE_FILE = 0;
+    const ID_CARD_FILE = 1;
+    const SELFIE_IMAGE_STRING = 2;
+    const ID_CARD_IMAGE_STRING = 3;
+    const LIVENESS_IMAGE_FILE = 4;
+    const ID_CARD_BACK_FILE = 5;
+    const LIVENESS_IMAGE_STRING = 6;
+    const ID_CARD_BACK_STRING = 7;
+}