+++ lib/JobType.php
@@ -3,8 +3,8 @@
 class JobType
 {
     const BIOMETRIC_KYC = 1;
-    const SMART_SELFIE_REGISTRATION = 2;
-    const SMART_SELFIE_AUTHENTICATION = 4;
+    const SMART_SELFIE_AUTHENTICATION = 2;
+    const SMART_SELFIE_REGISTRATION = 4;
     const BASIC_KYC = 5;
     const ENHANCED_KYC = 5;
     const DOCUMENT_VERIFICATION = 6;