+++ CHANGELOG.md
@@ -6,9 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [4.0.1] - 2022-12-12
 ### Added
 - Business verification to the iDApi class.
+- JobType, ImageType, and BusinessVerificationType enum classes
 - [examples](/examples) folder
 
 ## [4.0.0] - 2022-12-06